Second attempt
Been working on this problem for too long.  I have a working solution but it sucks if the target number is too big.
Learned a lot of valuable lessons from this....

I need to remember, now is the time for practice.  Ask questions and look things up. If I don't know the problem, I don't know it.
Give it my best shot and if I don't know it, that's fine, I'm learning.